fix(layout): fix page footer (#29)
* fix(layout): fix page footer links and theme switch accessibility
Improved the page footer by relocating hard-coded links to the 'footerLinks' constant, enhancing accessibility with an aria-label for the icon-only button, and optimizing the ThemeSwitchLabel placement by moving it to the RadioGroupItem. These changes enhance maintainability and user experience.

* fix(layout): delete unnecessary button components
The Button component was removed from the link items as it was unnecessary and causing styling and layout issues. This change simplifies the code and improves the overall layout consistency of the page footer.

Co-authored-by: Tika <179169711+tikatrsk@users.noreply.github.com>