feat(blog): add table of content to blog posts (#69)
* feat(blog): add table of content to blog posts
This commit introduces a table of contents feature for blog posts. It includes the addition of a table of contents to the post page, implementation of a new hook to observe the table of contents, creation of a new component to render the table of contents, and definition of a new type for the Velite table of contents. These enhancements improve navigation and readability for blog readers.

Co-authored-by: Tika <179169711+tikatrsk@users.noreply.github.com>